Rename code/locales to core/locale
