+++ frontend/src/core/locale/index.js
@@ -4,9 +4,9 @@ export { default as actions } from './actions';
 export { default as reducer } from './reducer';
 
 export type { Locale } from './actions';
-export type { LocalesState } from './reducer';
+export type { LocaleState } from './reducer';
 
 
 // Name of this module.
 // Used as the key to store this module's reducer.
-export const NAME: string = 'locales';
+export const NAME: string = 'locale';
