+++ frontend/src/core/locale/reducer.test.js
