+++ frontend/src/modules/entitydetails/components/EntityDetails.test.js
@@ -85,12 +85,8 @@ function createEntityDetailsWithStore() {
                 search: '?string=' + ENTITIES[0].pk,
             },
         },
-        locales: {
-            locales: {
-                'kg': {
-                    code: 'kg',
-                },
-            },
+        locale: {
+            code: 'kg',
         },
     };
     const store = createReduxStore(initialState);
