+++ frontend/src/rootReducer.js
@@ -5,7 +5,7 @@ import { combineReducers } from 'redux';
 import * as editor from 'core/editor';
 import * as entities from 'core/entities';
 import * as lightbox from 'core/lightbox';
-import * as locales from 'core/locales';
+import * as locale from 'core/locale';
 import * as l10n from 'core/l10n';
 import * as notification from 'core/notification';
 import * as plural from 'core/plural';
@@ -27,7 +27,7 @@ export default combineReducers({
     [editor.NAME]: editor.reducer,
     [entities.NAME]: entities.reducer,
     [lightbox.NAME]: lightbox.reducer,
-    [locales.NAME]: locales.reducer,
+    [locale.NAME]: locale.reducer,
     [l10n.NAME]: l10n.reducer,
     [notification.NAME]: notification.reducer,
     [plural.NAME]: plural.reducer,
